fix some common occurring errors
