+++ src/resources/emoji.json
@@ -3,23 +3,23 @@
   "madurai": "<:madurai:319586146499690496>",
   "naramon": "<:naramon:319586146478850048>",
   "zenurik": "<:zenurik:319586146197700610>",
-  "electricity": "<:electricity:321463957212626944>",
-  "cold": "<:cold:363136256659685377>",
-  "heat": "<:heat:363136256927858698>",
-  "toxin": "<:toxin:363136256626130946>",
-  "radiation": "<:radiation:363136256865075200>",
-  "viral": "<:viral:363136256747765763>",
+  "electricity": "<:electricity:604819783724171264>",
+  "cold": "<:cold:604819783892074496>",
+  "heat": "<:heat:604819783933886464>",
+  "toxin": "<:toxin:604819785053634610>",
+  "radiation": "<:radiation:604820111718744064>",
+  "viral": "<:viral:604821627091419156>",
   "gas": "<:gas:363136257045561344>",
-  "<DT_IMPACT>": "<:impact:363136256781189120>",
-  "impact": "<:impact:363136256781189120>",
-  "<DT_PUNCTURE>": "<:puncture:363136257129185280>",
-  "puncture": "<:puncture:363136257129185280>",
-  "<DT_SLASH>": "<:slash:363136256755892225>",
-  "slash": "<:slash:363136256755892225>",
+  "<DT_IMPACT>": "<:impact:604819300523442189>",
+  "impact": "<:impact:604819300523442189>",
+  "<DT_PUNCTURE>": "<:puncture:604819300288692226>",
+  "puncture": "<:puncture:604819300288692226>",
+  "<DT_SLASH>": "<:slash:604819299957211136>",
+  "slash": "<:slash:604819299957211136>",
   "koneski": "<:koneski:319586146483044352>",
   "unairu": "<:unairu:319586146453553162>",
-  "blast": "<:blast:363136256907149312>",
-  "corrosive": "<:corrosive:363136257288568832>",
-  "magnetic": "<:magnetic:363136420602445824>",
+  "blast": "<:blast:604820111756492810>",
+  "corrosive": "<:corrosive:604820111815081984>",
+  "magnetic": "<:magnetic:604820111949561895>",
   "umbra": "<:umbra:459116667255914496>"
 }
